reporegistry: handle distro ID strings with and without the dot
Rework the reporegistry and loading to of all repo configurations to
support distro ID strings with dot to separate minor version. This is
done in a backward compatible way, so that e.g. `rhel-89.json` and
`rhel-8.9.json` repo configs are considered to be for the same
`RHEL-8.9` distro.

The backward compatibility is achieved by always using the
`distroidstrparser` to first parse any external distro ID string and
then transforming the `distro.ID` back to its string representation
(which always uses dot to separate minor release version if set).

Rework unit tests to verify that repositories with and without a dot in
their filename override each other in case they get parsed to the same
`distro.ID`.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>