test: new script: generate build config quick
This new build config generator uses unresolved manifests to decide if
an image configuration has changed and should be rebuilt.

The process works as follows:
- Generate manifests without content in a directory "head"
- Create a temporary git worktree and check out the merge-base between
  HEAD and origin/main
- Generate manifests from the worktree without content in "merge-base"
- Diff the manifests in the two directories based on their *content*
- Generate a json file that lists the build configurations (distro,
  arch, image type, config name) of every build configuration that
  either changed or was introduced in the branch HEAD
- Upload the json file to S3, keyed by the HEAD commit ID

This file will be read by the gitlab runner to generate a pipeline for
test builds.

The idea is that we can generate manifests without resolving content and
be reasonably certain that any changes in these manifests reflect
changes in the real ones.
This is based on the assumption that the content doesn't change if the
address (URL) for the content is the same.  When gen-manifests is asked
to not resolve content, --packages=false, --commits=false, and/or
--containers=false, instead of depsolving, fetching the commit ID, or
resolving the container ID, it generates a checksum of each item based
on the provided info (name, url, etc). This is a safe alternative under
certain conditions:

packages: In our tests we use rpm repo snapshots.  That means that for a
given timestamped URL, the content will never change.  This isn't true
in the general case.  If we ever change a test repo to point to CDNs for
example, or anything that's not a pinned snapshot, then this assumption
falls apart.

ostree commits: For our tests, any build that requires an ostree commit
uses an ostree-container built by the manifest at the same commit.  It's
not really important what these images point to for this check: if their
manifest changes, but the corresponding ostree-container doesn't, it
means that something changed in the definition of the dependent manifest
(e.g., iot-qcow2, iot-installer, etc), so a rebuild will be needed.  If
a dependency changes (e.g., iot-container), we would start a rebuild of
that, and then automatically also rebuild any dependants, since the
content hash in their manifests will change in the real case.

containers: This is a bit stranger for cases where we embed containers.
Right now we use a couple of containers that we keep pinned in the
GitLab registry.  If we change one of those, this new method will never
detect the change because it doesn't try to resolve them.  We could
enable container resolves (it probably wont be that slow), but also we
almost never change those and when we do, we can just push them with a
new tag or name and update the test configs.  It might be a good idea to
refer to containers in blueprints by their image or manifest ID instead
of their tag.

Signed-off-by: Achilleas Koutsou <achilleas@koutsou.net>