manifest: use kickstart.Options on AnacondaInstallerISOTree
Remove all the kickstart-related properties from
AnacondaInstallerISOTree and add the new kickstart.Options struct.
Panic if it's not set and any of the stages require it, for example to
set the path of the file when a kickstart file is required for the
specific payload.

This commit has no effect on generated manifests.