manifest: use kickstart.Options on AnacondaInstaller
Remove the Users and Groups properties from AnacondaInstaller and add
the new kickstart.Options struct. Name it InteractiveDefaultsKickstart
to make it clear that it's not the same kind of kickstart file that we
create in the AnacondaInstallerISOTree but it only serves to set options
in the interactive default kickstart file. It currently only supports
Users and Groups.

This commit has no effect on generated manifests.