otk-gen-partiton-table: generate strictly internal representation for now
As a starting point generate internal vars from the partition
table. This means in the otk manifest one woudl write:
```
otk.define.defaults:
  filesystems:
    otk.external.osbuild.gen_partition_table:
      total_size: 10G
```
and it would get replaced with:
```
otk.define.defaults:
  filesystems:
    internal_partition_table:
      ...:
    kernel_opts_list:
      - foo
```
and then our "partition_table_make_osbuild_stage" would consume
the "internal_partition_table" and we would declare that those
are really (for now) just meant for the tools and otk manifests
should not try to do anything with them (beside passing to the
companion tool).

Later we could consider a nicer internal representation?