customizations/kargs: StringList() for kargs.Options
Method for returning a list of strings (unjoined).  Sometimes kernel
arguments need to be handled as a list of strings (for example in the
stage options for org.osbuild.ostree.deploy.container), so let's
implement that and then make the String() method a quick call to
strings.Join() on the output of that.