customizations/kargs: Update() for kargs.Options
Function that updates an instance of kargs.Options in-place with values
from another.  The exact behaviour of the function for each property
depends on the type and functionality of the property.
Values that are unset in newOpts do not affect the receiver.
String arrays, like Console, ModprobeBlacklist, and Extra, are
concatenated.