distro: add test_otk_distro
Add the skeleton of a test distribution that is built on otk yaml files.
Most methods return empty strings or nil values.  Some wont be used in
the test or will be implemented later, but are required to implement the
distro.Distro, distro.Arch, and distro.ImageType interfaces.

The distro creates a distro->arch->imagetype hierarchy based on the
contents of a filesystem tree.  It assumes a directory structure
`<distroname>/<archname>/<imagetypename>.yaml`.  For example:

fedora-39/
  x86_64/
    qcow2.yaml
    ami.yaml
  aarch64/
    qcow2.yaml