disk: add filesystem type to CreateMountpoint()
Support setting the filesystem type when calling CreateMountpoint() on a
PartitionTable, Btrfs volume, or LVM Volume Group.
For Btrfs volumes, only "btrfs" and the empty string are valid.

Calls to pt.CreateMountpoint() for "/boot" are always set to xfd for
backwards compatibility (at least for now).

In pt.createFilesystem(), when the fstype isn't specified, it
automatically inherits the filesystem type of the root filesystem.

This commit changes the behaviour of customizations in Fedora.  By
default until now, custom mountpoints were always xfs.  In RHEL and
CentOS this matched the root filesystem.  But in Fedora, the root
filesystem is ext4.  Therefore, after this commit, custom filesystms in
Fedora will default to ext4 instead of xfs.