disk: move vg and lv names up the stack
Make the VolumeGroup and LogicalVolume names arguments to
PartitionTable.ensureLVM().  Currently, they only take effect when the
root partition is not on LVM and gets converted.

This commit has no effect on generated manifests.

In particular, the name of the LogicalVolume, "root", becomes "rootlv"
after being sanitised by the lvname() function.