disk: don't sanitise lvname in CreateLogicalVolume()
The lvName passed to CreateLogicalVolume() was always assumed to be the
mountpoint for the filesystem, so it was always sanitised be a valid
LVM LogicalVolume name.  The lvname() function also appends the letters
'lv' to the sanitised mountpoint.

Since we intend to make the LogicalVolume name configurable, let's only
run mountpoints through the sanitiser function (lvname()) and not
externally set names.

This commit has no effect on generated manifests.