disk: create generalized entityPath
This is basically the same method as entityPathForMountpoint - a
DFS-search through a partition table tree. It takes a start and a
function. Every visited node (entity) is passed in such function. If the
function returns true, this means that the passed entity is the goal.

entityPathForMountpoint is now implemented using the new generalized
method.