disk: add bios and esp partitions based on boot mode
Add BIOS boot and ESP partitions based on boot mode.

Also, move the handling of the /boot partition before the plain
partition handling so we don't have to find where to insert it (after
BIOS boot and ESP but before plain partitions).  Instead, we build the
partition list in the correct order.

Update tests.