disk: move handling of the /boot partition
Move the handling of the /boot partition before the plain partition
handling so we don't have to find where to insert it (after BIOS boot
and ESP, which are going to be added in a later commit, but before
plain partitions).  Instead, we build the partition list sequentially as
needed.

Update tests.