blueprint: include minsize in partitioning customization
Add minsize to the top level of the partitioning customizations.
In osbuild-composer, this value comes from a command line option through
distro.ImageOptions.  The plan is to have that be a blueprint
customization that osbuild-composer will resolve when it's defined in
multiple places and pass it to images via the customization.