disk: relayout the partition table in NewCustomPartitionTable()
Run a relayout() on the partition table before returning it from
NewCustomPartitionTable() to set the start position and size for each
partition as well as the total partition table size.

Update tests.

Partition sizes in tests changed to MiB (from plain bytes) so that the
values make more sense when headers are added and values are rounded to
the grain size (1 MiB).