container: add Resolve() and ResolveAll() to resolver
Extend the Resolver interface to implement Resolve() and ResolveAll().
This is consistent with the API of the other resolvers.

The asyncResolver only implements stubs for these functions and panics
if they are called.  This resolver will be removed soon.

The blockingResolver implements Resolve() as its main resolver function.
The Add() function now wraps Resolve() and collects the resolved specs
in the results slice.
ResolveAll() calls Resolve() iteratively on all container source specs.

WIP: container resolver