osbuild: add Swap section for systemd.unit.create stage options
Add a Swap section and support for .swap systemd units.
This commit also adds a separate filename regex for unit files that
differs from the one we have for general files.  Specifically, it allows
for backslashes, which we will need to create .swap unit files.  The
filenames for these units must be the escaped path to the device or file
they control, which can contain escape characters (e.g. \x2d for
escaping a dash).

See also systemd.swap(5)