osbuild: define CommandLineOptions struct for osbuild
The osbuild.RunOSBuild() function arguments list is getting longer as we
expose more of the command line options from osbuild.  Put them all in a
single struct so we can expand it without expanding the number of args
and without changing the call signature.

This struct only holds `--` options that are passed directly to osbuild.
The manifest, as a required positional argument, is kept as the first
function argument.
Other options, like the env vars, are kept as function arguments to
separate them from the direct osbuild options.

This commit will require changes in all projects that import this
library.