manifest: mount root device rw and set rootflags when using mount units
The systemd-remount-fs service reads mount options from /etc/fstab for
the root partition (and /usr, if separate).  If there is no /etc/fstab,
the fs is not remounted, so a read-only mount will stay read-only on
boot.

Switch the "ro" command line argument to "rw" if we're not writing an
/etc/fstab file.
Also, add the root mount options as rootflags.