manifest: set boot_root to ESP mountpoint when using UKI
When we're using a UKI and we have a partition table defined, set the
boot_root option in the rpm stage to the ESP mountpoint.

The ESP mountpoint search code is extracted into a function for reuse.
Ideally, we'd only do this once in the serialize() function and pass it
wherever it's needed, but the current structure of the function makes
that a bit dirtier.  I find the current way, calling it twice, close to
where it's needed, is more readable.