manifest: add hmac and dir stages when needed
The hmac and dir stages are needed when the uki-direct package is older
than v25.3.

The 99-uki-uefi-setup.install script, prior to v25.3 of the uki-direct
package, would run `bootctl -p` to discover the ESP [1].  This doesn't
work in osbuild because the system isn't booted or live.  Since v25.3,
the install script respects the `$BOOT_ROOT` env var that we set in
osbuild during the org.osbuild.rpm stage.
The only things missing when the file isn't run are the .hmac file used
to verify the kernel during early boot in FIPS mode and the kernel
addons directory (at <ESP>/EFI/Linux/<kernel>.extra.d/).

[1] https://gitlab.com/kraxel/virt-firmware/-/commit/ca385db4f74a4d542455b9d40c91c8448c7be90c