distrodefs/rhel8: add {supported,required}_blueprint_options
Define the supported and required blueprint options for all the image
types in RHEL 8.  These are very similar to the ones added to the image
types in Fedora, with a few exceptions:
- Ignition is not supported for the edge-simplified-installer.
- RHEL distros define a "tar" image type that requires its own set of
  supported options.
- The "azure-eap7-rhui" image type does not support any customizations,
  even though it's a regular disk image.  It's only used internally and
  is never customized.

The checkOptionsRhel8() function has also been updated to match the
Fedora and RHEL 7 version, with modifications for the RHEL 8 image
types (type names, swap creation on aarch64).