distro: PoC blueprint options test
Proof-of-concept test for the supported blueprint options.
The idea for the test is to generate one blueprint for each supported
option and generate two manifests each time: one without the option and
one with.  The manifests should not be the same, since any supported
option should have an effect on the manifest.
The inverse should also be true: For each unsupported option, adding the
option should create a manifest identical to the one without the option.

Problems with the idea:
- Exceptions will have to be hardcoded in the test, which isn't a big
  issue.
- The options validator will have to be disabled for the unsupported
  test case.  This will need a bit of mocking work.
- Options with restricted valid values will need be encoded somehow.
- Invalid combinations of options will need to be managed.

Alternatives:
- Handwrite mappings from blueprint option name (path) to valid
  blueprint component.  For example:
    "customizations.kernel": {Name: "altkernel", Append: []string{"arg"}}
  using random values where appropriate, and use those instead of using
  reflection.  The downside is that this will require adding new
  elements to the map every time a new blueprint option is added, but
  that would probably not be too much work.
- Wait until we have the new blueprint and use the openapi schema to
  generate valid and invalid test cases.