distro: rename set_release_ver_var to set_releasever_var
The dnf variable name is `releasever` and it is more consistent with
other user-facing use cases where it is treated as a single word.  So
let's treat it internally as a single word as well.

This matches the new blueprint customization.

Similarly, rename the struct field from SetReleaseVerVar to
SetReleaseverVar to match.