distro: support combining dnf_config.set_releasever_var with options
Previously this was not supported because the
dnf_config.options.variables can set a releasever which would conflict
with the variable set by enabling dnf_config.set_releasever_var.
However, it's trivial to check for a conflicting variable of the same
name and make that the only error condition.

The tests have been rewritten to test all combinations.