test: add build_requests() function and fixture
This is a rewrite of the _test_cases() generator in
test_build_integration.py, written by Michael in
1c1a8d2c552d822c40d7f7498a2ab26be14e600d.

This makes each build request (distro, arch, image-type, config) its own
fixture.  This is similar to the parameters generated by _test_cases(),
but the objects can be scoped and reused between tests.

Like in the existing _test_cases() generator, the configurations are
generated dynamically using the gen-manifests command (with --dry-run).