test: add build_cache and cached_image fixtures
Two new pytest fixtures for handling the build cache and cached images.

The build_cache fixture populates a local tree with the info.json and
bib* files from the cache.  This fixture can be used to determine if an
image for a specific manifest exists in the cache.

The cached_image fixture combines the build_cache fixture with the
manifest to download the image that corresponds to that manifest, if
it's already cached.