test: add image fixture
A pytest fixture for retrieving a built image.

The fixture extends the cached_image fixture and returns the same
structure, but ensures that the "image-path" property is set (not None)
and points to the image that corresponds to the build request and
manifest.

If the image exists in the cache, this fixture simply returns the same
object as the cached_image fixture.  Otherwise, it builds it.